추가: 지도 -> 정보수정제안 (인텐트)
