+++ update-cafe/build.gradle.kts
@@ -36,5 +36,6 @@ dependencies {
     projects.designResource,
     projects.navigator,
     projects.domain,
+    projects.main,
   )
 }
