+++ update-cafe/src/main/kotlin/us/wedemy/eggeum/android/updatecafe/ui/SelectInfoCategoriesFragment.kt
@@ -29,6 +29,7 @@ class SelectInfoCategoriesFragment : BaseFragment<FragmentSelectInfoCategoriesBi
   private val viewModel by viewModels<SelectInfoCategoriesViewModel>()
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
     initListener()
     initObserver()
   }
