+++ update-cafe/src/main/kotlin/us/wedemy/eggeum/android/updatecafe/viewmodel/ProposeCafeInfoViewModel.kt
@@ -9,6 +9,7 @@
 
 package us.wedemy.eggeum.android.updatecafe.viewmodel
 
+import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -18,55 +19,85 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import us.wedemy.eggeum.android.domain.model.place.PlaceEntity
 import us.wedemy.eggeum.android.domain.model.place.ProductEntity
 import us.wedemy.eggeum.android.domain.usecase.GetPlaceUseCase
 import us.wedemy.eggeum.android.domain.usecase.UpsertPlaceUseCase
+import us.wedemy.eggeum.android.main.model.CafeDetailModel
+import us.wedemy.eggeum.android.updatecafe.ui.item.CafeInfoItem
 import us.wedemy.eggeum.android.updatecafe.ui.item.CafeMenuItem
 
 @HiltViewModel
 class ProposeCafeInfoViewModel @Inject constructor(
   private val getPlaceUseCase: GetPlaceUseCase,
   private val upsertPlaceUseCase: UpsertPlaceUseCase,
+  savedStateHandle: SavedStateHandle,
 ) : ViewModel() {
+  private val selectedCafeInfo: CafeDetailModel = savedStateHandle[CAFE_DETAIL_INFO]
+    ?: throw IllegalArgumentException("CAFE_DETAIL_INFO is missing from SavedStateHandle")
 
-  private val _cafeMenuList = MutableStateFlow(emptyList<CafeMenuItem>())
-  val cafeMenuList = _cafeMenuList.asStateFlow()
-
-  lateinit var placeBody: PlaceEntity
+  private val _navigateToUpsertEvent = MutableSharedFlow<Boolean>()
+  val navigateToUpsertEvent = _navigateToUpsertEvent.asSharedFlow()
 
   private val _showToastEvent = MutableSharedFlow<String>()
   val showToastEvent: SharedFlow<String> = _showToastEvent.asSharedFlow()
 
-  private val _cafeMenuItemMap = MutableStateFlow(emptyMap<String, CafeMenuItem>())
-  private val cafeMenuItemMap = _cafeMenuItemMap.asStateFlow()
-
   private val _getInitCall = MutableSharedFlow<Unit>(1)
   val getInitCall = _getInitCall.asSharedFlow()
 
-  private val _updatePlaceBodySuccess = MutableSharedFlow<Boolean>()
-  val updatePlaceBodySuccess = _updatePlaceBodySuccess.asSharedFlow()
+  private val _cafeMenuItemMap = MutableStateFlow(emptyMap<String, CafeMenuItem>())
+  private val cafeMenuItemMap = _cafeMenuItemMap.asStateFlow()
 
   fun setCafeMenuItemMap(cafeMenuItemMap: MutableMap<String, CafeMenuItem>) {
     _cafeMenuItemMap.value = cafeMenuItemMap
   }
 
+  lateinit var placeBody: PlaceEntity
+
+  private val _cafeMenuList = MutableStateFlow(emptyList<CafeMenuItem>())
+  val cafeMenuList = _cafeMenuList.asStateFlow()
+
+  private val _cafeInfo = MutableStateFlow(CafeInfoItem())
+  val cafeInfo = _cafeInfo.asStateFlow()
+
   init {
-    // TODO: 지도에서 placdId 받아와서 인자 넣기
-    getCafeMenuList(1)
+    getCafeMenuListAndInfo(selectedCafeInfo.id)
   }
 
-  private fun getCafeMenuList(placeId: Int) {
+  private fun getCafeMenuListAndInfo(placeId: Int) {
     viewModelScope.launch {
       val result = getPlaceUseCase(placeId)
       when {
         result.isSuccess && result.getOrNull() != null -> {
           placeBody = result.getOrNull()!!
-          placeBody.menu?.products?.let {
-            val cafeMenuItemList: MutableList<CafeMenuItem> = initializeCafeMenuItem(products = it)
-            updateCafeMenuList(cafeMenuItemList = cafeMenuItemList)
+          launch {
+            _cafeInfo.update { cafeInfo ->
+              cafeInfo.copy(
+                areaSize = placeBody.info?.areaSize,
+                blogUri = placeBody.info?.blogUri,
+                businessHours = placeBody.info?.businessHours,
+                existsSmokingArea = placeBody.info?.existsSmokingArea,
+                existsWifi = placeBody.info?.existsWifi,
+                instagramUri = placeBody.info?.instagramUri,
+                meetingRoomCount = placeBody.info?.meetingRoomCount,
+                mobileCharging = placeBody.info?.mobileCharging,
+                multiSeatCount = placeBody.info?.multiSeatCount,
+                parking = placeBody.info?.parking,
+                phone = placeBody.info?.phone,
+                restRoom = placeBody.info?.restRoom,
+                singleSeatCount = placeBody.info?.singleSeatCount,
+                websiteUri = placeBody.info?.websiteUri,
+              )
+            }
+          }
+          launch {
+            placeBody.menu?.products?.let {
+              val cafeMenuItemList: MutableList<CafeMenuItem> = initializeCafeMenuItem(products = it)
+              updateCafeMenuList(cafeMenuItemList = cafeMenuItemList)
+            }
           }
           _getInitCall.emit(Unit)
         }
@@ -117,8 +148,7 @@ class ProposeCafeInfoViewModel @Inject constructor(
       val result = upsertPlaceUseCase(placeBody.toUpsertPlaceEntity())
       when {
         result.isSuccess -> {
-          // TODO: 메뉴 수정을 완료했어요 fragment
-          _updatePlaceBodySuccess.emit(true)
+          _navigateToUpsertEvent.emit(true)
         }
         result.isFailure -> {
           val exception = result.exceptionOrNull()
@@ -129,9 +159,14 @@ class ProposeCafeInfoViewModel @Inject constructor(
     }
   }
 
+  // TODO: 에뮬레이터와 달리 실기기로 테스트 시, 해당 state value를 초기화 안해도 마지막 완료 화면으로 넘어가지 않음. (확인 필요)
   fun initializeUpdatePlaceBodySuccess() {
     viewModelScope.launch {
-      _updatePlaceBodySuccess.emit(false)
+      _navigateToUpsertEvent.emit(false)
     }
+  }
+
+  private companion object {
+    private const val CAFE_DETAIL_INFO = "cafe_detail_info"
   }
 }
